Oprava formatovania odpovede a uprava http stavovych kodov

+++ backend/app.js
@@ -43,16 +43,16 @@ function check_user(username,email){
 
 function check_token(token, res){
     if (tokens[token] == undefined) {
-        res.status(400).send("Invalid token")
-        return false
+        res.status(401).json({"message": "Unauthorized user or invalid token."});
+        return false;
     }
     return true
 }
 
 function check_token_and_id(token, id, res) {
     if(tokens[token] == undefined || id == undefined){
-        res.status(400).send("Invalid token or id");
-        return false
+        res.status(401).json({"message": "Unauthorized user or invalid token."});
+        return false;
     }
     return true
 }
@@ -93,7 +93,7 @@ app.get('/users/signUser', (req,res)=>{
         res.json({'message':'OK','token':t, 'shelter': data.shelter, 'email': data.email});
     })
     .catch((error)=>{
-            res.status(400).json({'message':'Invalid username or password'});
+            res.status(404).json({'message':'Invalid username or password'});
     })
 
 })
@@ -107,6 +107,14 @@ app.post('/users/register',(req,res)=>{
     console.log(username,password,email);
     // ak pouzivatel neexistuje
     if (check_user(username,email) == false){
+        if (username == undefined || email == undefined || password == undefined || shelter == undefined) {
+            res.status(404).json({'message':'Not all attributes provided'});
+            return;
+        }
+        if (typeof username != "string" || typeof email != "string" || typeof password != "string" || typeof shelter != "number") {
+            res.status(400).json({'message':'Not proper format of attributes.'});
+            return;
+        }
         db.one('INSERT INTO users(name,email,password,shelter) VALUES($1, $2, $3,$4) RETURNING id,shelter', [username, email, password,shelter])
         .then((data)=>{
             t = uuidv4();
@@ -118,11 +126,11 @@ app.post('/users/register',(req,res)=>{
         })
         .catch((error)=>{
             console.log(error)
-            res.status(400).json({'message':'Fail'});
+            res.status(404).json({'message':'Inserting data was not successful'});
         })
     }
     else{
-        res.status(400).send({'message':'User already exists'});
+        res.status(409).json({'message':'User already exists'});
     }
 })
 //nacitanie psov
@@ -158,7 +166,7 @@ app.get('/dogs/getAll', (req, res) => {
         })
         .catch((error) => {
             //res.json(error)
-            res.status(400).json({'message': 'Wrong request'})
+            res.status(404).json({'message': 'No data found'})
             return
         })
     }
@@ -185,7 +193,7 @@ app.get('/dogs/getAll', (req, res) => {
         .catch((error) => {
             res.json(error)
             console.log('error')
-            res.status(400).json({'message': 'Wrong request'})
+            res.status(404).json({'message': 'No data found'})
         })
     }
 })
@@ -200,6 +208,14 @@ app.get('/dogs/getDog', (req, res) => {
     userID = tokens[token]["id"];
     shelter = tokens[token]["shelter"];
     if (shelter == true) {
+        if (dog_id == undefined) {
+            res.status(404).json({'message':'Not all attributes provided'});
+            return;
+        }
+        if (typeof dog_id != "number") {
+            res.status(400).json({'message':'Not proper format of attributes.'});
+            return;
+        }
         db.one('SELECT * FROM dogs WHERE dogs.id = $1 AND dogs.shelter_id = $2', [dog_id, userID])
         .then((data) => {
             var raw_data = data.image_data
@@ -223,7 +239,7 @@ app.get('/dogs/getDog', (req, res) => {
         })
         .catch((error)=>{
             console.log(error)
-            res.status(400).json({'message':'Wrong request'})
+            res.status(404).json({'message':'No data found'})
         })
     }
     else {
@@ -249,7 +265,7 @@ app.get('/dogs/getDog', (req, res) => {
         })
         .catch((error)=>{
             console.log(error)
-            res.status(400).json({'message':'Wrong request'})
+            res.status(404).json({'message':'No data found'})
         })
     }
 })
@@ -280,7 +296,7 @@ app.get('/terms', (req, res) => {
         })
         .catch((error) => {
             console.log(error)
-            res.status(400).json({'message':'Wrong request'})
+            res.status(404).json({'message':'No data found'})
         })
     }
     else {
@@ -300,7 +316,7 @@ app.get('/terms', (req, res) => {
         })
         .catch((error) => {
             console.log(error)
-            res.status(400).json({'message':'Wrong request'})
+            res.status(404).json({'message':'No data found'})
         })
     }
 })
@@ -320,15 +336,25 @@ app.post('/dogs/addDog', (req, res) => {
     details = req.body.details;
     console.log(req.body);
     //photo = req.body.photo;
-    if (dog_name == undefined || breed == undefined || age == undefined || details == undefined || health == undefined || shelter == false) {
-        res.status(400).json({"message": "Bad params"});
+    if (dog_name == undefined || breed == undefined || age == undefined || details == undefined || health == undefined) {
+        res.status(404).json({"message": "Not all attributes provided"});
+        return;
+    }
+
+    if (shelter == false) {
+        res.status(401).json({"message": "Signed user is not a shelter"});
+        return;
+    }
+
+    if (typeof dog_name != "string" || typeof breed != "string" || typeof age != "number" || typeof details != "string" || typeof health != "string") {
+        res.status(400).json({'message':'Not proper format of attributes.'});
         return;
     }
 
     //db.one("INSERT INTO dogs (name, breed, age, details, image_location, shelter_id, health) VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING ID", [dog_name, breed, age, details, photo, userID, health])
     db.one("INSERT INTO dogs (name, breed, age, details, shelter_id, health) VALUES ($1, $2, $3, $4, $5, $6) RETURNING ID", [dog_name, breed, age, details, userID, health])
     .then((data) => res.status(200).json({"message":"OK", "id": data.id}))
-    .catch((error)=> res.status(400).json({"message": "Something went wrong"}))
+    .catch((error)=> res.status(400).json({"message": "No data found"}))
 })
 
 //uprava psa
@@ -345,15 +371,25 @@ app.put('/dogs/editDog', (req, res) => {
     details = req.body.details;
     shelter = tokens[token]["shelter"];
     userID = tokens[token]["id"];
-    if (dog_name == undefined || breed == undefined || age == undefined || details == undefined || health == undefined || dog_id == undefined || shelter == false) {
-        res.status(400).send("Bad params or user is not a shelter");
+    if (dog_name == undefined || breed == undefined || age == undefined || details == undefined || health == undefined || dog_id == undefined) {
+        res.status(404).json({"message": "Not all attributes provided"});
+        return;
+    }
+
+    if (shelter == false) {
+        res.status(401).json({"message": "Signed user is not a shelter"});
+        return;
+    }
+
+    if (typeof dog_name != "string" || typeof breed != "string" || typeof age != "number" || typeof details != "string" || typeof health != "string") {
+        res.status(400).json({'message':'Not proper format of attributes.'});
         return;
     }
 
     db.one("UPDATE dogs SET name = $1, breed = $2, age = $3, details = $4, health = $5  WHERE id = $6 and shelter_id=$7 RETURNING id", 
         [dog_name, breed, age, details, health, dog_id,userID])
-    .then((data) => res.send("OK"))
-    .catch((error)=> res.status(400).send("Bad request"))
+    .then((data) => res.status(200).json({"message": "OK"}))
+    .catch((error)=> res.status(404).json({"message": "No data to be updated"}))
 })
 
 //vymazanie psa
@@ -361,16 +397,31 @@ app.delete('/dogs/deleteDog', (req, res) => {
     token = req.query.token;
     dog_id = req.query.dog_id;
     if (!check_token_and_id(token, dog_id, res))
-        return
+        return;
+
+    if (dog_id == undefined) {
+        res.status(404).json({"message": "Not all attributes provided"});
+        return;
+    }
+
+    if (shelter == false) {
+        res.status(401).json({"message": "Signed user is not a shelter"});
+        return;
+    }
+
+    if (typeof dog_id != "number") {
+        res.status(400).json({'message':'Not proper format of attributes.'});
+        return;
+    }
 
     userID = tokens[token]["id"];
-    db.any("DELETE FROM dogs WHERE dogs.shelter_id = $1 AND dogs.id = $2;", [userID, dog_id])
+    db.one("DELETE FROM dogs WHERE dogs.shelter_id = $1 AND dogs.id = $2;", [userID, dog_id])
     .then((data) => {
         db.any("DELETE FROM terms WHERE terms.dog_id = $2; DELETE FROM forms WHERE forms.dog_id = $2", [userID, dog_id])
-        .then((data) => res.send("OK"))
-        .catch((error) => res.status(400).send("Bad request"))
+        .then((data) => res.status(200).json({"message": "OK"}))
+        .catch((error) => res.status(404).json({"message": "No data to be deleted"}))
     })
-    .catch((error) => res.status(400).send("Bad request this"))
+    .catch((error) => res.status(404).json({"message": "No data to be deleted"}))
 })
 
 // vytvorenie formulara
@@ -384,8 +435,18 @@ app.post('/forms/create', (req,res)=>{
     dog_id = req.body.dog_id;
     type = req.body.type;
     details = req.body.details;
-    if(details == undefined || dog_id == undefined || type==undefined || shelter == true){
-        res.status(400).send("Bad params");
+    if(details == undefined || dog_id == undefined || type==undefined){
+        res.status(404).json({"message": "Not all attributes provided"});
+        return;
+    }
+
+    if (shelter == true) {
+        res.status(401).json({"message": "Signed user is a shelter"});
+        return;
+    }
+
+    if (typeof details != "string" || typeof dog_id != "number" || typeof type != "number") {
+        res.status(400).json({'message':'Not proper format of attributes.'});
         return;
     }
 
@@ -397,24 +458,24 @@ app.post('/forms/create', (req,res)=>{
             if(term_id == undefined){
                 db.any("DELETE FROM forms WHERE id=$1",[data.id])
                 .then((data)=>{
-                    res.status(400).json({"message": "Something went wrong"});
+                    res.status(400).json({"message": "Not all attributes provided"});
                 })
                 .catch((error)=>{
-                    res.status(400).json({"message": "Something went wrong"});
+                    res.status(404).json({"message": "No data to be deleted and not all attributes provided"});
                 }) 
             }
             db.one("UPDATE terms SET form_id=$1 WHERE terms.id=$2 RETURNING id",[data.id,term_id])
             .then((data)=>{
-                res.json({"message":"OK"})
+                res.status(200).json({"message":"OK"})
             })
             .catch((error)=>{
-                res.status(400).json({"message": "Something went wrong"});
+                res.status(404).json({"message": "No data to be updated"});
             })
         }
         res.status(200).json({"message": "OK"});
     })
     .catch((error)=>{
-        res.status(400).json({"message": "Something went wrong"});
+        res.status(400).json({"message": "No data found"});
     })
 })
 //načítanie detailu formulára
@@ -423,6 +484,16 @@ app.get('/forms/detail',(req,res)=>{
     token = req.query.token;
     if (!check_token_and_id(token, form_id, res))
         return;
+
+    if (form_id == undefined) {
+        res.status(404).json({"message": "Not all attributes provided"});
+        return;
+    }
+
+    if (typeof form_id != "number") {
+        res.status(400).json({'message':'Not proper format of attributes.'});
+        return;
+    }
     
     userID = tokens[token]["id"];
     db.one("SELECT * FROM forms WHERE id=$1 AND (forms.user_id=$2 OR forms.dog_id IN (SELECT dogs.id FROM dogs WHERE shelter_id = $2))",[form_id, userID])
@@ -436,11 +507,11 @@ app.get('/forms/detail',(req,res)=>{
             //pridane po milestone 2
             "finished": data.finished
         }
-        res.json(result)
+        res.json(result);
     })
     .catch((error)=>{
         console.log(error)
-        res.status(400).send("Wrong request")
+        res.status(404).json({"message" : "No data found"})
     })
 })
 app.get('/forms/getAll',(req,res)=>{
@@ -464,7 +535,7 @@ app.get('/forms/getAll',(req,res)=>{
     })
     .catch((error)=>{
         console.log(error)
-        res.status(400).send("No forms for this user")
+        res.status(404).json({"message": "No data found"})
     })
 })
 // editovanie formulara
@@ -478,16 +549,26 @@ app.put('/forms/edit',(req,res)=>{
     details = req.body.details;
     finished = req.body.finished;
     if (formId == undefined || details == undefined || finished == undefined) {
-        res.status(400).send("Bad params");
+        res.status(404).json({"message": "Not all attributes provided"});
+        return;
+    }
+
+    if (shelter == true) {
+        res.status(401).json({"message": "Signed user is a shelter"});
+        return;
+    }
+
+    if (typeof formId != "number" || typeof details != "string" || typeof finished != "string") {
+        res.status(400).json({'message':'Not proper format of attributes.'});
         return;
     }
 
     db.one("UPDATE forms SET details = $1, finished = $2 WHERE id=$3 AND user_id = $4 RETURNING id", [details,finished,formId, userID])
     .then((data)=>{
-        res.json({"message":"OK"})
+        res.send(200).json({"message":"OK"})
     })
     .catch((error)=>{
-        res.status(400).json({"message":"Bad request"})
+        res.status(404).json({"message":"No data to be updated"})
     })
 })
 // vymazanie formulara
@@ -497,15 +578,25 @@ app.delete('/forms/delete',(req,res)=>{
     if (!check_token_and_id(token, form_id, res))
         return
 
+    if (form_id == undefined) {
+        res.status(404).json({"message": "Not all attributes provided"});
+        return;
+    }
+
+    if (typeof form_id != "number") {
+        res.status(400).json({'message':'Not proper format of attributes.'});
+        return;
+    }
+
     userID = tokens[token]["id"];
     db.one("DELETE from forms WHERE id=$2 AND (forms.user_id=$1 OR forms.dog_id IN (SELECT dogs.id FROM dogs WHERE shelter_id = $1)) RETURNING dog_id",[userID,form_id])
     .then((data)=>{
         db.any("DELETE FROM terms WHERE dog_id = $1 and form_id = $2", [data.dog_id, form_id])
-        .then((data) => res.json({"message":"OK"}))
-        .catch((error)=> res.status(400).json({"message":"Bad request"}))
+        .then((data) => res.status(200).json({"message":"OK"}))
+        .catch((error)=> res.status(404).json({"message":"No data to be deleted"}))
     })
     .catch((error)=>{
-        res.status(400).send("Bad request");
+        res.status(404).json({"message": "No data to be deleted"});
     })
 })
 // vytvorenie terminov pre psa
@@ -513,6 +604,22 @@ app.post('/terms/create',(req,res)=>{
     dog_id = req.query.dog_id;
     if (!check_token_and_id(req.query.token, dog_id, res))
         return;
+
+    if (dog_id == undefined) {
+        res.status(404).json({"message": "Not all attributes provided"});
+        return;
+    }
+
+    if (shelter == false) {
+        res.status(401).json({"message": "Signed user is not a shelter"});
+        return;
+    }
+
+    if (typeof dog_id != "number") {
+        res.status(400).json({'message':'Not proper format of attributes.'});
+        return;
+    }
+
     token = req.query.token;
     userID = tokens[token]["id"];
     //ziskam posledny formular pre psa
@@ -525,7 +632,7 @@ app.post('/terms/create',(req,res)=>{
     .catch((error)=>{
         insert_terms(dog_id,new Date())
     })
-    res.send("OK")
+    res.status(200).json({"message": "OK"})
 
 })
 function insert_terms(dog_id, time){
@@ -551,17 +658,22 @@ app.put('/terms/update',(req,res)=>{
 
     free = req.body.free
     if (free == undefined) {
-        res.status(400).send("Bad params");
+        res.status(404).json({"message": "Not all attributes provided"});
+        return;
+    }
+
+    if (typeof free != "string") {
+        res.status(400).json({'message':'Not proper format of attributes.'});
         return;
     }
 
     user_id = tokens[token]["id"]
     db.one("UPDATE terms SET free=$1, user_id=$2 WHERE id=$3 RETURNING id",[free,user_id,term_id])
     .then((data)=>{
-        res.send("OK")
+        res.status(200).json({"message": "OK"})
     })
     .catch((error)=>{
-        res.status(400).send("Bad request")
+        res.status(404).json({"message": "No data to be updated"})
     })
 })
 
@@ -571,6 +683,16 @@ app.get('/image', (req, res) => {
     dog_id = req.query.dog_id;
     if (!check_token_and_id(token, dog_id, res))
         return
+
+    if (dog_id == undefined) {
+        res.status(404).json({"message": "Not all attributes provided"});
+        return;
+    }
+
+    if (typeof dog_id != "number") {
+        res.status(400).json({'message':'Not proper format of attributes.'});
+        return;
+    }
 
     userID = tokens[token]["id"]
     shelter = tokens[token]["shelter"]
@@ -591,7 +713,7 @@ app.get('/image', (req, res) => {
         })
         .catch((error)=>{
             console.log(error)
-            res.status(400).json({'message':'Wrong request'})
+            res.status(404).json({'message':'No data found'})
         })
     }
     else {
@@ -613,7 +735,7 @@ app.get('/image', (req, res) => {
         })
         .catch((error)=>{
             console.log(error)
-            res.status(400).json({'message':'Wrong request'})
+            res.status(404).json({'message':'No data found'})
         })
     }
     
@@ -627,10 +749,21 @@ app.post('/image/insert', upload.single('file'), (req, res) => {
     if (!check_token_and_id(token, dog_id, res))
         return;
 
-    if (req.file == undefined) {
-        res.status(400).json({"message":"Bad params"});
+    if (req.file == undefined || dog_id == undefined) {
+        res.status(404).json({"message": "Not all attributes provided"});
         return;
     }
+
+    if (shelter == false) {
+        res.status(401).json({"message": "Signed user is not a shelter"});
+        return;
+    }
+
+    if (typeof dog_id != "number") {
+        res.status(400).json({'message':'Not proper format of attributes.'});
+        return;
+    }
+
     type = req.file.mimetype;
     image_name = req.file.originalname;
     image_data = req.file.buffer;
@@ -643,20 +776,19 @@ app.post('/image/insert', upload.single('file'), (req, res) => {
     })
     .catch((error)=>{
         console.log(error)
-        res.status(400).json({'message':'Wrong request'})
+        res.status(404).json({'message':'No data to be updated'})
     })
 })
 
 //odhlasenie pouzivatela
 app.get('/users/logout', (req, res) => {
     token = req.query.token;
-    if (!check_token(token, res)){
-        res.status(400).json({"message": "Wrong request"});
+    if (!check_token(token, res))
         return
-    }
+    
     tokens[token] = undefined;
     console.log("Pohodka");
-    res.json({"message": "OK"});
+    res.status(200).json({"message": "OK"});
 })
 
 app.listen(port, () => {
